chore: extract submitContractReport route handler
